+++ src/samplers/scheduler/linux/mod.rs
@@ -1,2 +1,1 @@
-#[cfg(feature = "bpf")]
 mod runqueue;